fixed more level 4 warnings

+++ Tests/TestsRosicAndRapt/Source/rosic_tests/PortedFromRSLib/Examples/ModalExamples.cpp
@@ -134,7 +134,8 @@ std::vector<double> applyCombWeighting(std::vector<double> v, std::vector<double
 std::vector<double> pseudoHarmonicRatios12TET(int numPartials) // only 0..20
 {
   double tmp[21];
-  long double s = pow(2.0, 1.0/12.0); // basis
+  //long double s = pow(2.0, 1.0/12.0); // basis
+  double s = pow(2.0, 1.0/12.0); // basis
 
                          // #    ratio
   tmp[0]  = pow(s,  0);  //  1    1.0
@@ -169,8 +170,8 @@ std::vector<double> ratios12TET(int numPartials)
 {  
   double tmp[21];
   std::vector<double> r(numPartials);
-  long double s = pow(2.0, 1.0/12.0); // basis
-
+  //long double s = pow(2.0, 1.0/12.0); // basis
+  double s = pow(2.0, 1.0/12.0); // basis
                          //  #    ratio
   tmp[0]  = pow(s,  0);  //  1    1.0
   tmp[1]  = pow(s, 12);  //  2    2.0
